build: adjust uv management and reduce configuration
