+++ src/posit/__init__.py
@@ -1,7 +1,3 @@
 """The Posit SDK."""
 
-from . import connect  # noqa
-
-__all__ = [
-    "connect"
-]
+from . import connect as connect
