+++ src/posit/connect/__init__.py
@@ -1,1 +1,1 @@
-from .client import Client  # noqa
+from .client import Client as Client
