+++ src/posit/connect/metrics/__init__.py
@@ -1,18 +1,1 @@
-"""Metric resources."""
-
-from .. import resources
-from .usage import Usage
-
-
-class Metrics(resources.Resources):
-    """Metrics resource.
-
-    Attributes
-    ----------
-    usage: Usage
-        Usage resource.
-    """
-
-    @property
-    def usage(self) -> Usage:
-        return Usage(self.params)
+from .metrics import Metrics as Metrics
