+++ vars.mk
@@ -28,7 +28,11 @@ endif
 
 NETLIFY_SITE_ID ?= 5cea1f56-7935-4387-975a-18a7905d15ee
 
+ifneq ($(shell command -v uv 2>/dev/null),)
+PYTHON := python
+else
 PYTHON := python3
+endif
 
 ifneq ($(shell command -v uv 2>/dev/null),)
 PIP := uv pip
